requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
bidict==0.22.1
binance==0.3
binance-connector==3.5.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dateparser==1.2.0
flask==3.0.1
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6
flask-sqlalchemy==3.1.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
load-dotenv==0.1.0
MarkupSafe==2.1.5
multidict==6.0.5
mysql-connector-python==8.3.0
pycryptodome==3.20.0
PyJWT==2.8.0
PyMySQL==1.1.0
python-binance==1.0.19
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.8.2
python-socketio==5.11.0
pytz==2024.1
regex==2023.12.25
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.25
typing-extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
ujson==5.9.0
urllib3==2.2.0
websocket-client==1.7.0
websockets==12.0
werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
